Move JsoupParser tests to package JsoupParserTests
